requirements.txt
keras
viam-sdk==0.50.0
typing-extensions

